+++ main.tf
@@ -2,16 +2,6 @@ data "aws_caller_identity" "peer" {
   provider = aws.peer
 }
 
-data "aws_vpc" "this" {
-  id = var.vpc_id
-}
-
-data "aws_vpc" "peer" {
-  provider = aws.peer
-
-  id = var.peer_vpc_id
-}
-
 resource "aws_vpc_peering_connection" "this" {
   peer_owner_id = data.aws_caller_identity.peer.account_id
   peer_vpc_id   = var.peer_vpc_id
@@ -27,30 +17,22 @@ resource "aws_vpc_peering_connection_accepter" "this" {
   tags                      = var.peer_tags
 }
 
-resource "aws_route" "public" {
-  count = length(var.public_route_tables)
+resource "aws_route" "this" {
+  for_each = { for route in var.routes : route.name => route if route.provider == "aws" }
 
-  route_table_id            = var.public_route_tables[count.index]
-  destination_cidr_block    = data.aws_vpc.peer.cidr_block
-  vpc_peering_connection_id = aws_vpc_peering_connection.this.id
-}
-
-resource "aws_route" "private" {
-  count = length(var.private_route_tables)
-
-  route_table_id            = var.private_route_tables[count.index]
-  destination_cidr_block    = data.aws_vpc.peer.cidr_block
-  vpc_peering_connection_id = aws_vpc_peering_connection.this.id
+  route_table_id              = each.value.route_table_id
+  destination_cidr_block      = each.value.destination_cidr_block
+  destination_ipv6_cidr_block = each.value.destination_ipv6_cidr_block
+  vpc_peering_connection_id   = aws_vpc_peering_connection.this.id
 }
 
 resource "aws_route" "peer" {
-  count = length(var.peer_route_tables)
+  for_each = { for route in var.routes : route.name => route if route.provider == "aws.peer" }
 
   provider = aws.peer
 
-  route_table_id            = var.peer_route_tables[count.index]
-  destination_cidr_block    = data.aws_vpc.this.cidr_block
-  vpc_peering_connection_id = aws_vpc_peering_connection_accepter.this.id
-
-  depends_on = [aws_vpc_peering_connection_accepter.this]
+  route_table_id              = each.value.route_table_id
+  destination_cidr_block      = each.value.destination_cidr_block
+  destination_ipv6_cidr_block = each.value.destination_ipv6_cidr_block
+  vpc_peering_connection_id   = aws_vpc_peering_connection_accepter.this.id
 }