+++ tests/generated_route_tables/main.tf
@@ -6,10 +6,9 @@ module "generated_route_tables" {
     aws.peer = aws
   }
 
-  vpc_id               = module.vpc_pcx_requester.vpc_id
-  peer_vpc_id          = module.vpc_pcx_accepter.vpc_id
-  public_route_tables  = module.vpc_pcx_requester.public_route_table_ids
-  private_route_tables = module.vpc_pcx_requester.private_route_table_ids
+  vpc_id      = module.vpc_pcx_requester.vpc_id
+  peer_vpc_id = module.vpc_pcx_accepter.vpc_id
+  routes      = concat(local.routes, local.routes_peer)
 
   tags = {
     Name = "tardigrade-pcx-${random_string.this.result}"
@@ -33,8 +32,11 @@ module "vpc_pcx_requester" {
 module "vpc_pcx_accepter" {
   source = "github.com/terraform-aws-modules/terraform-aws-vpc?ref=v3.14.0"
 
-  name = "tardigrade-pcx-vpc_pcx_accepter-${random_string.this.result}"
-  cidr = "10.1.0.0/16"
+  name            = "tardigrade-pcx-accepter-${random_string.this.result}"
+  cidr            = "10.1.0.0/16"
+  azs             = ["us-east-1a", "us-east-1b", "us-east-1c"]
+  private_subnets = ["10.1.1.0/24", "10.1.2.0/24", "10.1.3.0/24"]
+  public_subnets  = ["10.1.101.0/24", "10.1.102.0/24", "10.1.103.0/24"]
 }
 
 resource "random_string" "this" {
@@ -42,4 +44,26 @@ resource "random_string" "this" {
   number  = false
   special = false
   upper   = false
+}
+
+locals {
+  routes = [for i, rt in concat(module.vpc_pcx_requester.public_route_table_ids, module.vpc_pcx_requester.private_route_table_ids) :
+    {
+      name                        = "route-${i}"
+      provider                    = "aws"
+      route_table_id              = rt
+      destination_cidr_block      = "10.1.0.0/16"
+      destination_ipv6_cidr_block = null
+    }
+  ]
+
+  routes_peer = [for i, rt in concat(module.vpc_pcx_accepter.public_route_table_ids, module.vpc_pcx_accepter.private_route_table_ids) :
+    {
+      name                        = "route-${i}"
+      provider                    = "aws.peer"
+      route_table_id              = rt
+      destination_cidr_block      = "10.0.0.0/16"
+      destination_ipv6_cidr_block = null
+    }
+  ]
 }